Varying terrain: Refactor all classes. Pass noise, seed_poits, etc into geography. Time creation of components.
